refactor(behavioral,agent,modules): add explicit invalid detail matching
Add BPListener detailMatch support with default valid semantics and update

listener matching to honor explicit invalid mode without changing source

matching behavior.

Replace production/server-agent notSchema usage with explicit listeners

that set detailMatch to invalid in create-agent, use-installer, and

server module guards.

Flatten the websocket server core module to src/modules/server.ts,

update exports/imports, and add tests covering invalid listener paths and

valid flow preservation. Also update behavioral-core skill references to

use explicit schemas/detailMatch guidance for server-agent patterns.
