+++ src/modules.ts
@@ -1,4 +1,1 @@
-export * from './modules/server/server-module.constants.ts'
-export * from './modules/server/server-module.schemas.ts'
-export * from './modules/server/server-module.ts'
-export type * from './modules/server/server-module.types.ts'
+export * from './modules/server.ts'
